requirement.txt
requests
pycryptodome
pytest
pytest-html

